Fix bar chart circular dependencies
