+++ src/components/MultiSeriesBarChart/components/StackedBarGroup/index.ts
@@ -1,1 +1,1 @@
-export {StackedBarGroup, StackedBarGroupProps} from './StackedBarGroup';
+export {StackedBarGroup} from './StackedBarGroup';
