+++ src/components/MultiSeriesBarChart/components/index.ts
@@ -1,2 +1,2 @@
 export {BarGroup} from './BarGroup';
-export {StackedBarGroup, StackedBarGroupProps} from './StackedBarGroup';
+export {StackedBarGroup} from './StackedBarGroup';
